requirements.txt
# LangChain Core
langchain 
langchain-core

# Groq/OpenAI-Compatible API Integration
langchain-groq
openai  
tiktoken 

# HuggingFace Integration
langchain-huggingface 
transformers
huggingface-hub
sentence-transformers

# Google Gemini Integration
langchain-google-genai
google-generativeai

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn 

# Framework
torch
torchvision
torchaudio

# User Interface
streamlit

# Visualization
grandalf